refactor(commons): split into files todo api functions
