+++ packages/commons/src/api/index.ts
@@ -1,4 +1,4 @@
-import * as todo from './todo';
+import { todo } from './todo';
 import * as timeline from './timeline';
 import * as users from './users';
