+++ packages/commons/src/api/todo/get-todo.ts
@@ -1,0 +1,12 @@
+import { TODO_URL } from "../../utils/constants";
+
+export const getTodo = async (id: string) => {
+  const resp = await fetch(`${TODO_URL}/find/${id}`);
+  const data = await resp.json();
+
+  if (resp.status === 404) {
+    throw new Error(data.message);
+  }
+
+  return data;
+};
