+++ packages/commons/src/utils/constants.ts
@@ -3,3 +3,7 @@ const API_URL = import.meta.env.VITE_TODO_API;
 export const TODO_URL = `${API_URL}/todos`;
 export const TIMELINE_URL = `${API_URL}/timeline`;
 export const USERS_URL = `${API_URL}/users`;
+
+export const headers = {
+  'Content-Type': 'application/json',
+};
